# openapi.yaml
...
openapi: 3.1.0
info:
  title: SMIRK API
  version: "1.0.0"
  description: Route inventory generated from concrete Express route declarations in server.ts and src/team-routes.ts. Middleware app.use declarations are excluded from paths.
servers:
  - url: https://smirkcalls.com
    description: Production
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    WorkspaceBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Workspace session or invite token
    ProvisioningBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Provisioning token
paths:
...
  "/api/integrations/crm":
    get:
      operationId: "GET_api_integrations_crm"
      tags:
        - "integrations"
      summary: "GET /api/integrations/crm"
      x-source: "src/routes/integrations-routes.ts:75"
      x-express-path: "/api/integrations/crm"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
  "/api/integrations/crm/test":
    post:
      operationId: "POST_api_integrations_crm_test"
      tags:
        - "integrations"
      summary: "POST /api/integrations/crm/test"
      x-source: "src/routes/integrations-routes.ts:85"
      x-express-path: "/api/integrations/crm/test"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
  "/api/integrations/webhook":
    get:
      operationId: "GET_api_integrations_webhook"
      tags:
        - "integrations"
      summary: "GET /api/integrations/webhook"
      x-source: "src/routes/integrations-routes.ts:39"
      x-express-path: "/api/integrations/webhook"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
...
  "/api/integrations/webhook/test":
    post:
      operationId: "POST_api_integrations_webhook_test"
      tags:
        - "integrations"
      summary: "POST /api/integrations/webhook/test"
      x-source: "src/routes/integrations-routes.ts:50"
      x-express-path: "/api/integrations/webhook/test"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
...
  "/api/mcp":
    get:
      operationId: "GET_api_mcp"
      tags:
        - "mcp"
      summary: "GET /api/mcp"
      x-source: "src/routes/integrations-routes.ts:144"
      x-express-path: "/api/mcp"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
    post:
      operationId: "POST_api_mcp"
      tags:
        - "mcp"
      summary: "POST /api/mcp"
      x-source: "src/routes/integrations-routes.ts:149"
      x-express-path: "/api/mcp"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
  "/api/mcp-servers":
    get:
      operationId: "GET_api_mcp_servers"
      tags:
        - "mcp-servers"
      summary: "GET /api/mcp-servers"
      x-source: "src/routes/integrations-routes.ts:190"
      x-express-path: "/api/mcp-servers"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
    post:
      operationId: "POST_api_mcp_servers"
      tags:
        - "mcp-servers"
      summary: "POST /api/mcp-servers"
      x-source: "src/routes/integrations-routes.ts:193"
      x-express-path: "/api/mcp-servers"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
  "/api/mcp-servers/{id}":
    delete:
      operationId: "DELETE_api_mcp_servers_id"
      tags:
        - "mcp-servers"
      summary: "DELETE /api/mcp-servers/:id"
      x-source: "src/routes/integrations-routes.ts:199"
      x-express-path: "/api/mcp-servers/:id"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
    put:
      operationId: "PUT_api_mcp_servers_id"
      tags:
        - "mcp-servers"
      summary: "PUT /api/mcp-servers/:id"
      x-source: "src/routes/integrations-routes.ts:196"
      x-express-path: "/api/mcp-servers/:id"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
  "/api/mcp/{id}":
    delete:
      operationId: "DELETE_api_mcp_id"
      tags:
        - "mcp"
      summary: "DELETE /api/mcp/:id"
      x-source: "src/routes/integrations-routes.ts:165"
      x-express-path: "/api/mcp/:id"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
    put:
      operationId: "PUT_api_mcp_id"
      tags:
        - "mcp"
      summary: "PUT /api/mcp/:id"
      x-source: "src/routes/integrations-routes.ts:158"
      x-express-path: "/api/mcp/:id"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
  "/api/mcp/{id}/test":
    post:
      operationId: "POST_api_mcp_id_test"
      tags:
        - "mcp"
      summary: "POST /api/mcp/:id/test"
      x-source: "src/routes/integrations-routes.ts:170"
      x-express-path: "/api/mcp/:id/test"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
...
  "/api/plugin-tools":
    get:
      operationId: "GET_api_plugin_tools"
      tags:
        - "plugin-tools"
      summary: "GET /api/plugin-tools"
      x-source: "src/routes/integrations-routes.ts:177"
      x-express-path: "/api/plugin-tools"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
    post:
      operationId: "POST_api_plugin_tools"
      tags:
        - "plugin-tools"
      summary: "POST /api/plugin-tools"
      x-source: "src/routes/integrations-routes.ts:180"
      x-express-path: "/api/plugin-tools"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
  "/api/plugin-tools/{id}":
    delete:
      operationId: "DELETE_api_plugin_tools_id"
      tags:
        - "plugin-tools"
      summary: "DELETE /api/plugin-tools/:id"
      x-source: "src/routes/integrations-routes.ts:186"
      x-express-path: "/api/plugin-tools/:id"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
    put:
      operationId: "PUT_api_plugin_tools_id"
      tags:
        - "plugin-tools"
      summary: "PUT /api/plugin-tools/:id"
      x-source: "src/routes/integrations-routes.ts:183"
      x-express-path: "/api/plugin-tools/:id"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
...
  "/api/tools":
    get:
      operationId: "GET_api_tools"
      tags:
        - "tools"
      summary: "GET /api/tools"
      x-source: "src/routes/integrations-routes.ts:111"
      x-express-path: "/api/tools"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
    post:
      operationId: "POST_api_tools"
      tags:
        - "tools"
      summary: "POST /api/tools"
      x-source: "src/routes/integrations-routes.ts:116"
      x-express-path: "/api/tools"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
  "/api/tools/{id}":
    delete:
      operationId: "DELETE_api_tools_id"
      tags:
        - "tools"
      summary: "DELETE /api/tools/:id"
      x-source: "src/routes/integrations-routes.ts:132"
      x-express-path: "/api/tools/:id"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
    put:
      operationId: "PUT_api_tools_id"
      tags:
        - "tools"
      summary: "PUT /api/tools/:id"
      x-source: "src/routes/integrations-routes.ts:125"
      x-express-path: "/api/tools/:id"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
  "/api/tools/{id}/test":
    post:
      operationId: "POST_api_tools_id_test"
      tags:
        - "tools"
      summary: "POST /api/tools/:id/test"
      x-source: "src/routes/integrations-routes.ts:137"
      x-express-path: "/api/tools/:id/test"
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Success response. Body shape is implemented in the route handler.
        "400":
          description: Bad request or validation failure.
        "401":
          description: Authentication required or invalid credentials.
        "500":
          description: Server error.
...